setup and start scripts corrected

+++ load_map.sh
@@ -2,8 +2,14 @@
 
 . moses.config
 
-docker stop mosesdb
-docker rm mosesdb
+if docker ps -a | grep -wq mosesdb;
+then
+	docker stop mosesdb
+	docker rm mosesdb
+fi
 
 docker run -p 5432:5432 -e 'POSTGRES_PASSWORD='$postgres_password -d --name mosesdb mosesdb_server:latest
-docker run -v $map_dir:/data --network="host" --name mosesdb_loader mosesdb_loader:latest
+
+sleep 5s
+
+docker run -v $map_dir:/data --rm --network="host" --name mosesdb_loader mosesdb_loader:latest
